Add Google login endpoint to AuthController
A new POST endpoint `user-login-google` has been added to the `AuthController` class in the `XuongMayBE.API.Controllers` namespace. This endpoint allows users to log in using Google. The `RefreshToken` method takes a `UserLoginGoogleRequest` object and returns an `IActionResult`. It calls the `_userService.UserLoginGoogle` method with the provided request and returns the result. If an exception occurs, it catches the exception and returns a `BadRequest` with an `ApiErrorResult` containing the exception message.