Add initial seeding for MembershipPackages in database
Introduced a mechanism in DependencyInjection.cs to check for existing
MembershipPackages in the database. If none are found, a predefined
list of MembershipPackage objects is initialized and added to the
database. This ensures that the application has predefined packages
available for users, enhancing the functionality and user experience
of the BabyCare API.